# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "mementos-backend",
        "description": "Documentation de l'API du projet Mementos",
        "contact": {
            "name": "Support Team",
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/revision/{themeId}": {
            "post": {
                "tags": [
                    "Revision"
                ],
                "summary": "Créer une révision pour un thème",
                "description": "Ajoute toutes les cartes d'un thème spécifique aux révisions de l'utilisateur connecté, à condition que le thème soit public ou appartienne à l'utilisateur.",
                "operationId": "createRevision",
                "parameters": [
                    {
                        "name": "themeId",
                        "in": "path",
                        "description": "L'identifiant du thème pour lequel créer une révision",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Révision créée avec succès, renvoie les détails du thème révisé.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Theme"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès refusé si le thème est privé et n'appartient pas à l'utilisateur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Accès refusé. Le thème et privé."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Thème déjà révisé par l'utilisateur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Thème déjà révisé."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur lors de la tentative d'ajout du thème aux révisions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erreur lors de l'ajout du thème à vos révisions: message d'erreur"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Revision"
                ],
                "summary": "Supprimer un thème des révisions de l'utilisateur connecté",
                "description": "Supprime toutes les révisions associées à un thème spécifique pour l'utilisateur connecté.",
                "operationId": "deleteThemeFromRevision",
                "parameters": [
                    {
                        "name": "themeId",
                        "in": "path",
                        "description": "Identifiant du thème à supprimer des révisions",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thème supprimé des révisions avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Thème supprimé des révisions avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur lors de la suppression du thème des révisions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erreur lors de la suppression du thème des révisions: <error message>"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/revision": {
            "put": {
                "tags": [
                    "Revision"
                ],
                "summary": "Mettre à jour le niveau d'une carte révisée",
                "operationId": "updateRevision",
                "requestBody": {
                    "description": "Données nécessaires pour mettre à jour la révision",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id",
                                    "is_correct"
                                ],
                                "properties": {
                                    "id": {
                                        "description": "Identifiant de la carte à réviser",
                                        "type": "integer",
                                        "example": 111
                                    },
                                    "is_correct": {
                                        "description": "Indique si la réponse était correcte",
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Révision mise à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur lors de la mise à jour de la révision",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erreur lors de la mise à jour de la révision: détail de l'erreur"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/revision/deleteAll": {
            "delete": {
                "tags": [
                    "Revision"
                ],
                "summary": "Supprimer toutes les révisions de l'utilisateur connecté",
                "description": "Effectue une suppression physique de toutes les révisions associées à l'utilisateur connecté.",
                "operationId": "deleteAllRevision",
                "responses": {
                    "200": {
                        "description": "Toutes les révisions ont été supprimées avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Toutes les révisions ont été supprimées avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur lors de la suppression des révisions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erreur lors de la suppression des révisions: <error message>"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}